fix: resolve comment 2

+++ migrations/sqls/20240303065352-add-product-table-up.sql
@@ -2,6 +2,6 @@ CREATE TABLE
     products (
         id VARCHAR(50) PRIMARY KEY,
         name VARCHAR(100) NOT NULL,
-        price integer NOT NULL,
+        price NUMERIC (10,2) NOT NULL,
         category VARCHAR(100)
     );
